🐛 Correct refresh behavior
Before this patch, we were not refreshing data from ecowatt once HA had
started. It was due to a mixup between `schedule_interval` and
`update_interval` 🤦.

This patch fixes this bug and avoids initial request to RTE API at
startup if we can spare it. Goal is to avoid hitting API limit (one call
per 15 minutes) when restarting.

Fixes #3

Change-Id: I2fdbc252d7874ff9ac5bc7bc4e5e00d763e2c108