🐛 Adapt to 2023.6
I should really learn not to use internal APIs :)

HA/core removes the async keyword on a method called "async" 🤦 leading
to #61.

This patch fixes it at the expense of not restoring "extra" attributes
(such as "generation time", "period start", "period end" and "level
code") which is very sad.

Fix #61

Change-Id: I3aade86daaa46cf83eb49a0e90ccb83c3dbeba9f